split up IDeserializerHandler.cs
