+++ src/LionWeb-CSharp/core/M1/IDeserializerHandler.cs
@@ -17,9 +17,7 @@
 
 namespace LionWeb.Core.M1;
 
-using M2;
 using M3;
-using Serialization;
 
 /// <summary>
 /// Callbacks to customize an <see cref="IDeserializer"/>'s behaviour in non-regular situations.
@@ -47,8 +45,8 @@ public interface IDeserializerHandler
     /// <param name="node">Currently deserialized node with id <paramref name="nodeId"/>.</param>
     /// <returns>Replacement node id to use for <paramref name="node"/>, or <c>null</c> to skip <paramref name="node"/>.</returns>
     /// <remarks>For both <paramref name="existingNode"/> and <paramref name="node"/>, only node id and properties are populated -- no other features.</remarks>
-    /// <remarks>if returned replacement node id is not unique, deserializer keeps calling this method, might lead to an infinite loop </remarks>
-    string? DuplicateNodeId(CompressedId nodeId, IReadableNode existingNode, INode node);
+    /// <remarks>If returned replacement node id is not unique, deserializer keeps calling this method, might lead to an infinite loop.</remarks>
+    string? DuplicateNodeId(CompressedId nodeId, IReadableNode existingNode, IReadableNode node);
 
     /// <summary>
     /// Cannot resolve <paramref name="metaPointer"/>, but know about at least one language
@@ -108,7 +106,7 @@ public interface IDeserializerHandler
     /// <param name="annotation">Invalid annotation used for <paramref name="node"/>.</param>
     /// <param name="node">Node that wants to have <paramref name="annotation"/> as annotation.</param>
     /// <returns>Replacement annotation node to use, or <c>null</c> to skip annotation <paramref name="annotation"/>.</returns>
-    IWritableNode? InvalidAnnotation(IReadableNode annotation, IReadableNode node);
+    IWritableNode? InvalidAnnotation(IReadableNode annotation, IReadableNode? node);
 
     /// <summary>
     /// Adding <paramref name="containedNode"/> as containment or annotation to <paramref name="parent"/> would create a containment cycle.
@@ -226,340 +224,4 @@ public interface IDeserializerHandler
     void InvalidAnnotationParent(IReadableNode annotation, IReadableNode? parent);
 
     #endregion
-}
-
-public static class DeserializerHandlerSelectOtherLanguageVersion
-{
-    /// <inheritdoc cref="SelectVersion{T}(LionWeb.Core.M1.CompressedMetaPointer,System.Collections.Generic.List{LionWeb.Core.M3.Language},Comparer{Language})"/>
-    public static T? SelectVersion<T>(CompressedMetaPointer metaPointer, List<Language> languages)
-        where T : class, IKeyed =>
-        SelectVersion<T>(metaPointer, languages, Comparer<Language>.Create(DefaultLanguageComparer));
-
-    /// <summary>
-    /// Chooses the <typeparamref name="T"/> with the same key as <paramref name="metaPointer"/>
-    /// from the <paramref name="languages">language</paramref> with the
-    /// <paramref name="languageComparer">first</paramref> version.
-    /// </summary>
-    /// <param name="metaPointer">Unresolvable meta-pointer.</param>
-    /// <param name="languages">Languages with same key as <paramref name="metaPointer"/>.</param>
-    /// <param name="languageComparer">Comparer to select the preferred language.
-    /// Make sure to invert the result for choosing the latest version.</param>
-    /// <typeparam name="T">Kind of language element we're looking for.</typeparam>
-    public static T? SelectVersion<T>(CompressedMetaPointer metaPointer, List<Language> languages,
-        Comparer<Language> languageComparer)
-        where T : class, IKeyed
-    {
-        IEnumerable<IKeyed> keyed = typeof(T) switch
-        {
-            { } f when f == typeof(Feature) => languages
-                .SelectMany(l => l.Entities.OfType<Classifier>())
-                .SelectMany(c => c.Features),
-            { } l when l == typeof(EnumerationLiteral) => languages
-                .SelectMany(l => l.Entities.OfType<Enumeration>())
-                .SelectMany(e => e.Literals),
-            _ => languages
-                .SelectMany(l => l.Entities.OfType<T>())
-        };
-
-        var candidates = keyed
-            .Where(k => CompressedId.Create(k.Key, false).Equals(metaPointer.Key))
-            .OrderBy(f => f.GetLanguage(), languageComparer);
-
-        return candidates
-            .FirstOrDefault() as T;
-    }
-
-    public static int DefaultLanguageComparer(Language a, Language b)
-    {
-        if (Version.TryParse(a.Version, out Version? aVersion) && Version.TryParse(b.Version, out Version? bVersion))
-        {
-            return -aVersion.CompareTo(bVersion);
-        }
-
-        return -string.Compare(a.Version, b.Version, StringComparison.Ordinal);
-    }
-}
-
-public class DeserializerExceptionHandler : IDeserializerHandler
-{
-    /// <inheritdoc />
-    public virtual Classifier? UnknownClassifier(CompressedMetaPointer classifier, CompressedId id) =>
-        throw new UnsupportedClassifierException(classifier, $"On node with id={id}: ");
-
-    /// <inheritdoc />
-    public virtual string? DuplicateNodeId(CompressedId nodeId, IReadableNode existingNode, INode node) =>
-        throw new DeserializerException($"Duplicate node with id={existingNode.GetId()}");
-
-    /// <inheritdoc />
-    public virtual T? SelectVersion<T>(CompressedMetaPointer metaPointer, List<Language> languages)
-        where T : class, IKeyed =>
-        // TODO think about correct handling
-        // throw new DeserializerException($"Unknown meta-pointer {metaPointer}");
-        null;
-
-    #region features
-
-    /// <inheritdoc />
-    public virtual Feature? UnknownFeature<TFeature>(CompressedMetaPointer feature,
-        Classifier classifier,
-        IReadableNode node) where TFeature : class, Feature =>
-        throw new UnknownFeatureException(classifier, feature, $"On node with id={node.GetId()}:");
-
-    /// <inheritdoc />
-    public virtual Feature? InvalidFeature<TFeature>(CompressedMetaPointer feature,
-        Classifier classifier,
-        IReadableNode node) where TFeature : class, Feature =>
-        throw new UnknownFeatureException(classifier, feature, $"On node with id={node.GetId()}:");
-
-    /// <inheritdoc />
-    public virtual List<T>? InvalidLinkValue<T>(List<T> value, Feature link, IReadableNode node)
-        where T : class, IReadableNode =>
-        throw new InvalidValueException(link, value);
-
-    /// <inheritdoc />
-    public virtual IWritableNode? InvalidAnnotation(IReadableNode annotation, IReadableNode node) =>
-        throw new DeserializerException(
-            $"On node with id={node.GetId()}: unsuitable annotation {annotation}");
-
-    /// <inheritdoc />
-    public virtual IWritableNode? CircularContainment(IReadableNode containedNode, IReadableNode parent) =>
-        throw new DeserializerException(
-            $"On node with id={parent.GetId()}: adding {containedNode.GetId()} as child/annotation would result in circular containment.");
-
-    /// <inheritdoc />
-    public virtual bool DuplicateContainment(IReadableNode containedNode, IReadableNode newParent,
-        IReadableNode existingParent) =>
-        throw new DeserializerException(
-            $"On node with id={containedNode.GetId()}: already has parent {existingParent.GetId()}, but also child/annotation of {newParent.GetId()}.");
-
-    #endregion
-
-    #region properties
-
-    /// <inheritdoc />
-    public virtual Enum? UnknownEnumerationLiteral(string key, Enumeration enumeration,
-        Feature property, IReadableNode node) =>
-        throw new DeserializerException(
-            $"On node with id={node.GetId()}: unknown enumeration literal for enumeration {enumeration} with key {key}");
-
-    /// <inheritdoc />
-    public virtual object? UnknownDatatype(Feature property, string? value, IReadableNode node) =>
-        throw new DeserializerException(
-            $"On node with id={node.GetId()}: unknown property type {property /*.Type*/} with value {value}");
-
-    /// <inheritdoc />
-    public virtual object? InvalidPropertyValue<TValue>(string? value, Feature property, CompressedId nodeId) =>
-        throw new DeserializerException(
-            $"On node with id={nodeId}: invalid property value '{value}' for property {property}");
-
-    #endregion
-
-    #region unresolveable nodes
-
-    /// <inheritdoc />
-    public virtual IWritableNode? UnresolvableChild(CompressedId childId, Feature containment, IReadableNode node) =>
-        throw new DeserializerException($"On node with id={node.GetId()}: couldn't find child with id={childId}");
-
-    /// <inheritdoc />
-    public virtual IReadableNode? UnresolvableReferenceTarget(CompressedId? targetId,
-        string? resolveInfo,
-        Feature reference,
-        IReadableNode node) =>
-        throw new DeserializerException(
-            $"On node with id={node.GetId()}: couldn't find reference target with id={targetId}");
-
-    /// <inheritdoc />
-    public virtual IWritableNode? UnresolvableAnnotation(CompressedId annotationId, IReadableNode node) =>
-        throw new DeserializerException(
-            $"On node with id={node.GetId()}: couldn't find annotation with id={annotationId}");
-
-    #endregion
-
-    #region language deserializer
-
-    /// <inheritdoc />
-    public virtual void InvalidContainment(IReadableNode node) =>
-        throw new UnsupportedClassifierException(node.GetClassifier().ToMetaPointer(),
-            $"On node with id={node.GetId()}:");
-
-    /// <inheritdoc />
-    public virtual void InvalidReference(IReadableNode node) =>
-        throw new UnsupportedClassifierException(node.GetClassifier().ToMetaPointer(),
-            $"On node with id={node.GetId()}:");
-
-    /// <inheritdoc />
-    public virtual void InvalidAnnotationParent(IReadableNode annotation, IReadableNode? parent) =>
-        throw new DeserializerException(
-            $"Cannot attach annotation {annotation} to its parent with id={parent?.GetId()}.");
-
-    /// <inheritdoc />
-    public virtual bool SkipDeserializingDependentNode(CompressedId id) =>
-        throw new DeserializerException(
-            $"Skip deserializing {id} because dependentLanguages contains node with same id");
-
-    #endregion
-}
-
-public class DeserializerException(string? message) : LionWebExceptionBase(message);
-
-public class DeserializerIgnoringHandler : IDeserializerHandler
-{
-    /// <inheritdoc />
-    public virtual Classifier? UnknownClassifier(CompressedMetaPointer classifier, CompressedId id)
-    {
-        LogMessage($"On node with id={id}: couldn't find specified classifier {classifier} - skipping.");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public string? DuplicateNodeId(CompressedId nodeId, IReadableNode existingNode, INode node)
-    {
-        LogMessage($"Duplicate node with id={existingNode.GetId()}");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public virtual T? SelectVersion<T>(CompressedMetaPointer metaPointer, List<Language> languages)
-        where T : class, IKeyed
-    {
-        LogMessage($"Unknown meta-pointer {metaPointer}");
-        return null;
-    }
-
-    #region features
-
-    /// <inheritdoc />
-    public virtual Feature? UnknownFeature<TFeature>(CompressedMetaPointer feature,
-        Classifier classifier,
-        IReadableNode node) where TFeature : class, Feature
-    {
-        LogMessage(
-            $"On node with id={node.GetId()}: couldn't find specified feature {feature} - leaving this feature unset.");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public List<T>? InvalidLinkValue<T>(List<T> value, Feature link, IReadableNode node) where T : class, IReadableNode
-    {
-        LogMessage($"On node with id={node.GetId()}: invalid link value {value} for link {link} - skipping");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public Feature? InvalidFeature<TFeature>(CompressedMetaPointer feature,
-        Classifier classifier,
-        IReadableNode node) where TFeature : class, Feature
-    {
-        LogMessage($"On node with id={node.GetId()}: wrong type of feature {feature} - leaving this feature unset.");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public IWritableNode? InvalidAnnotation(IReadableNode annotation, IReadableNode node)
-    {
-        LogMessage($"On node with id={node?.GetId()}: unsuitable annotation {annotation} - skipping.");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public IWritableNode? CircularContainment(IReadableNode containedNode, IReadableNode parent)
-    {
-        LogMessage(
-            $"On node with id={parent.GetId()}: adding {containedNode.GetId()} as child/annotation would result in circular containment - skipping");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public bool DuplicateContainment(IReadableNode containedNode, IReadableNode newParent,
-        IReadableNode existingParent)
-    {
-        LogMessage(
-            $"On node with id={containedNode.GetId()}: already has parent {existingParent.GetId()}, but also child/annotation of {newParent.GetId()} - keeping it in place.");
-        return false;
-    }
-
-    #endregion
-
-    #region properties
-
-    /// <inheritdoc />
-    public Enum? UnknownEnumerationLiteral(string key, Enumeration enumeration, Feature property, IReadableNode node)
-    {
-        LogMessage(
-            $"On node with id={node.GetId()}: unknown enumeration literal for enumeration {enumeration} with key {key} - skipping");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public object? UnknownDatatype(Feature property, string? value, IReadableNode node)
-    {
-        LogMessage(
-            $"On node with id={node.GetId()}: unknown datatype {property /*.Type*/} with value {value} - skipping");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public object? InvalidPropertyValue<TValue>(string? value, Feature property, CompressedId nodeId)
-    {
-        LogMessage($"On node with id={nodeId}: invalid property value {value} for property {property} - skipping");
-        return null;
-    }
-
-    #endregion
-
-    #region unresolveable nodes
-
-    /// <inheritdoc />
-    public virtual IWritableNode? UnresolvableChild(CompressedId childId, Feature containment, IReadableNode node)
-    {
-        LogMessage($"On node with id={node.GetId()}: couldn't find child with id={childId} - skipping.");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public virtual IReadableNode? UnresolvableReferenceTarget(CompressedId? targetId,
-        string? resolveInfo,
-        Feature reference,
-        IReadableNode node)
-    {
-        LogMessage($"On node with id={node.GetId()}: couldn't find reference target with id={targetId} - skipping.");
-        return null;
-    }
-
-    /// <inheritdoc />
-    public virtual IWritableNode? UnresolvableAnnotation(CompressedId annotationId, IReadableNode node)
-    {
-        LogMessage($"On node with id={node.GetId()}: couldn't find annotation with id={annotationId} - skipping.");
-        return null;
-    }
-
-    #endregion
-
-    #region language deserializer
-
-    /// <inheritdoc />
-    public void InvalidContainment(IReadableNode node) =>
-        LogMessage($"installing containments in node of meta-concept {node.GetType().Name} not implemented");
-
-    /// <inheritdoc />
-    public void InvalidReference(IReadableNode node) =>
-        LogMessage($"installing references in node of meta-concept {node.GetType().Name} not implemented");
-
-
-    /// <inheritdoc />
-    public void InvalidAnnotationParent(IReadableNode annotation, IReadableNode? parent) =>
-        LogMessage($"Cannot attach annotation {annotation} to its parent with id={parent?.GetId()}.");
-
-
-    /// <inheritdoc />
-    public bool SkipDeserializingDependentNode(CompressedId id)
-    {
-        LogMessage($"Skip deserializing {id} because dependent nodes contains node with same id");
-        return true;
-    }
-
-    #endregion
-
-    protected virtual void LogMessage(string format) =>
-        Console.WriteLine(format);
 }
